Update DealDTO, findByRenter and Vendor for Booking and Deal

+++ src/main/java/org/avengers/capstone/hostelrenting/dto/deal/DealDTOFull.java
@@ -11,5 +11,5 @@ public class DealDTOFull {
     private HostelTypeDTO type;
     private RenterDTO renter;
     private VendorDTO vendor;
-    private long offeredPrice;
+    private float offeredPrice;
 }
